# model/model.go
// Code generated by github.com/99designs/gqlgen, DO NOT EDIT.

package model

import (
	item "github.com/kautsarady/adindopustaka/model/item"
)

type Book struct {
	ID          int             `json:"id,omitempty"`
	Title       string          `json:"title,omitempty"`
	Authors     []item.Author   `json:"authors,omitempty"`
	Categories  []item.Category `json:"categories,omitempty"`
	ImageURL    string          `json:"image_url,omitempty"`
	GramedURL   string          `json:"gramed_url,omitempty"`
	Tags        []item.Tag      `json:"tags,omitempty"`
	Description string          `json:"description,omitempty"`
}